Add perms to commands

+++ index.js
@@ -25,8 +25,12 @@ client.on("message", async (msg) => {
   }
 
   if (message.startsWith("!addidea")) {
-    const idea = message.substr(message.indexOf(" ") + 1);
-    addIdea(msg, idea);
+    if (msg.member.roles.cache.find((r) => r.name === "Officers")) {
+      const idea = message.substr(message.indexOf(" ") + 1);
+      addIdea(msg, idea);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!addtodo")) {
@@ -36,7 +40,11 @@ client.on("message", async (msg) => {
   }
 
   if (message.startsWith("!ideas")) {
-    getIdeas(msg, dsc);
+    if (msg.member.roles.cache.find((r) => r.name === "Officers")) {
+      getIdeas(msg, dsc);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!todo")) {
@@ -45,8 +53,12 @@ client.on("message", async (msg) => {
   }
 
   if (message.startsWith("!removeidea")) {
-    const idea = message.substr(message.indexOf(" ") + 1);
-    deleteIdea(msg, idea);
+    if (msg.member.roles.cache.find((r) => r.name === "Officers")) {
+      const idea = message.substr(message.indexOf(" ") + 1);
+      deleteIdea(msg, idea);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!removetodo")) {
